refactor: make optimized gas manager the default and use alchemyEstimateFeesPerGas
- Deleted the non-optimized alchemyGasManagerHooks (ERC-7677 flow)
- Renamed alchemyGasAndPaymasterAndDataHooks to alchemyGasManagerHooks
- Gas manager now uses single alchemy_requestGasAndPaymasterAndData call by default
- Updated fee estimation to use alchemyEstimateFeesPerGas from smart-accounts package
- Removed manual rundler_maxPriorityFeePerGas calls in favor of typed client
- Updated all tests and imports to use the new API
- Regenerated documentation with the updated exports

BREAKING CHANGE: alchemyGasAndPaymasterAndDataHooks has been removed,
use alchemyGasManagerHooks instead which now uses the optimized flow by default
