+++ docs/docs.yml
@@ -405,8 +405,6 @@ navigation:
                     path: wallets/pages/reference/account-kit/infra/functions/alchemyEnhancedApiActions.mdx
                   - page: alchemyFeeEstimator
                     path: wallets/pages/reference/account-kit/infra/functions/alchemyFeeEstimator.mdx
-                  - page: alchemyGasAndPaymasterAndDataHooks
-                    path: wallets/pages/reference/account-kit/infra/functions/alchemyGasAndPaymasterAndDataHooks.mdx
                   - page: alchemyGasAndPaymasterAndDataMiddleware
                     path: wallets/pages/reference/account-kit/infra/functions/alchemyGasAndPaymasterAndDataMiddleware.mdx
                   - page: alchemyGasManagerHooks
