--- docs/pages/reference/account-kit/infra/functions/alchemyGasAndPaymasterAndDataHooks.mdx
@@ -1,52 +1,0 @@
----
-# This file is autogenerated
-title: alchemyGasAndPaymasterAndDataHooks
-description: Overview of the alchemyGasAndPaymasterAndDataHooks method
-slug: wallets/reference/account-kit/infra/functions/alchemyGasAndPaymasterAndDataHooks
----
-
-Adapts Alchemy's optimized gas+paymaster middleware to viem's new interface.
-Uses the non-standard `alchemy_requestGasAndPaymasterAndData` flow that combines
-gas estimation and paymaster sponsorship in a single RPC call.
-
-This implementation caches the result from the first call to avoid duplicate RPC calls
-when viem calls both getPaymasterStubData and getPaymasterData.
-
-## Import
-
-```ts
-import { alchemyGasAndPaymasterAndDataHooks } from "@account-kit/infra";
-```
-
-## Usage
-
-```ts
-import { createBundlerClient } from "viem/account-abstraction";
-import { alchemyGasAndPaymasterAndDataHooks } from "@account-kit/infra";
-
-const bundler = createBundlerClient({
-  transport: http("https://eth-sepolia.g.alchemy.com/v2/your-api-key"),
-  chain: sepolia,
-  account,
-  ...alchemyGasAndPaymasterAndDataHooks("your-policy-id"),
-});
-```
-
-## Parameters
-
-### policyId
-
-`string | string[]`
-
-- The policy ID(s) for Alchemy's gas manager
-
-### policyToken
-
-`PolicyToken`
-
-- Optional ERC-20 token configuration
-
-## Returns
-
-`{ paymaster: Function, userOperation: { estimateFeesPerGas: Function } }`
-Hooks for createBundlerClient
